[#71] refactored
