+++ collect_app/src/main/java/org/bcss/collect/android/jobs/CollectJobCreator.java
@@ -24,6 +24,8 @@ import com.evernote.android.job.JobCreator;
 
 import org.bcss.collect.naxa.jobs.DailyNotificationJob;
 import org.bcss.collect.naxa.jobs.LocalNotificationJob;
+import org.bcss.collect.naxa.jobs.MonthlyNotificationJob;
+import org.bcss.collect.naxa.jobs.WeeklyNotificationJob;
 import org.odk.collect.android.tasks.ServerPollingJob;
 
 public class CollectJobCreator implements JobCreator {
@@ -40,6 +42,10 @@ public class CollectJobCreator implements JobCreator {
                 return new LocalNotificationJob();
             case DailyNotificationJob.TAG:
                 return new DailyNotificationJob();
+            case WeeklyNotificationJob.TAG:
+                return new WeeklyNotificationJob();
+            case MonthlyNotificationJob.TAG:
+                return new MonthlyNotificationJob();
 
             default:
                 return null;
