+++ collect_app/src/main/java/org/bcss/collect/naxa/jobs/alarms/AlarmBroadcastReceiver.java
@@ -1,0 +1,29 @@
+package org.bcss.collect.naxa.jobs.alarms;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+import org.bcss.collect.android.R;
+import org.bcss.collect.android.application.Collect;
+import org.bcss.collect.naxa.common.FieldSightNotificationUtils;
+import org.bcss.collect.naxa.scheduled.data.ScheduleForm;
+import org.bcss.collect.naxa.scheduled.data.ScheduledFormsLocalSource;
+
+import java.util.List;
+
+public class AlarmBroadcastReceiver extends BroadcastReceiver {
+
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        List<ScheduleForm> scheduleForms = ScheduledFormsLocalSource.getInstance().getDailyForms();
+        String title = "Daily Reminder";
+        String message = Collect.getInstance().getString(R.string.msg_form_reminder_daily, scheduleForms.size());
+
+        FieldSightNotificationUtils.getINSTANCE().notifyNormal(title, message);
+    }
+
+    public static void setRepeatingEveryWeek(){
+
+    }
+}
