+++ collect_app/src/main/java/org/bcss/collect/naxa/preferences/SettingsFragment.java
@@ -1,20 +1,12 @@
 package org.bcss.collect.naxa.preferences;
 
 import android.app.TimePickerDialog;
-import android.content.Context;
 import android.content.SharedPreferences;
-import android.content.res.TypedArray;
-import android.graphics.Color;
-import android.graphics.drawable.ColorDrawable;
-import android.os.Build;
 import android.os.Bundle;
 import android.preference.Preference;
 import android.preference.PreferenceFragment;
 import android.preference.SwitchPreference;
 import android.support.v7.app.AlertDialog;
-import android.text.format.DateFormat;
-import android.util.AttributeSet;
-import android.view.Window;
 import android.widget.TimePicker;
 
 import com.evernote.android.job.JobManager;
@@ -23,10 +15,9 @@ import org.bcss.collect.android.BuildConfig;
 import org.bcss.collect.android.R;
 import org.bcss.collect.naxa.jobs.DailyNotificationJob;
 import org.bcss.collect.naxa.jobs.LocalNotificationJob;
-import org.odk.collect.android.utilities.ThemeUtils;
+import org.bcss.collect.naxa.jobs.MonthlyNotificationJob;
+import org.bcss.collect.naxa.jobs.WeeklyNotificationJob;
 
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
@@ -36,6 +27,8 @@ import timber.log.Timber;
 
 import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_SAMPLE;
 import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_SWITCH_DAILY;
+import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_SWITCH_MONTHLY;
+import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_SWITCH_WEEKLY;
 import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_TIME_DAILY;
 import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_TIME_MONTHLY;
 import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_TIME_WEEKLY;
@@ -44,8 +37,6 @@ import static org.bcss.collect.naxa.preferences.SettingsKeys.KEY_NOTIFICATION_TI
 public class SettingsFragment extends PreferenceFragment implements Preference.OnPreferenceClickListener, SharedPreferences.OnSharedPreferenceChangeListener {
 
 
-    private final boolean NOT_SYNCING = false;
-    ThemeUtils themeUtils = new ThemeUtils(SettingsFragment.this.getActivity());
     CustomTimePickerDialog customTimePickerDialog = null;
     private SwitchPreference switchPreferenceMonth, switchPreferenceDaily, switchPreferenceWeek;
     String[] weeks = {"Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"};
@@ -58,7 +49,9 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
         findPreference(KEY_NOTIFICATION_TIME_DAILY).setOnPreferenceClickListener(this);
         findPreference(KEY_NOTIFICATION_TIME_WEEKLY).setOnPreferenceClickListener(this);
         findPreference(KEY_NOTIFICATION_TIME_MONTHLY).setOnPreferenceClickListener(this);
+
         findPreference(KEY_NOTIFICATION_SAMPLE).setOnPreferenceClickListener(this);
+
         setupSharedPreferences();
         setupNotificationToggle();
         setupUpdateButton();
@@ -72,6 +65,8 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
             }
         }, Integer.valueOf(dailyTime[0]), Integer.valueOf(dailyTime[1]));
 
+        getPreferenceScreen().removePreference(findPreference(KEY_NOTIFICATION_TIME_WEEKLY));
+        getPreferenceScreen().removePreference(findPreference(KEY_NOTIFICATION_TIME_MONTHLY));
     }
 
     private void save(String key, String value) {
@@ -96,6 +91,7 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
         switchPreferenceWeek = (SwitchPreference) findPreference("switch_notification_weekly");
         switchPreferenceMonth = (SwitchPreference) findPreference("switch_notification_monthly");
 
+
         switchPreferenceDaily.setSummaryOff(getString(R.string.msg_no_longer_notifcation_receiced));
         switchPreferenceWeek.setSummaryOff(getString(R.string.msg_no_longer_notifcation_receiced));
         switchPreferenceMonth.setSummaryOff(getString(R.string.msg_no_longer_notifcation_receiced));
@@ -103,6 +99,7 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
         switchPreferenceDaily.setSummaryOn(getString(R.string.msg_will_receive_notifications));
         switchPreferenceWeek.setSummaryOn(getString(R.string.msg_will_receive_notifications));
         switchPreferenceMonth.setSummaryOn(getString(R.string.msg_will_receive_notifications));
+
     }
 
 
@@ -151,20 +148,31 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
             case KEY_NOTIFICATION_TIME_DAILY:
                 String time = String.valueOf(SettingsSharedPreferences.getInstance().get(KEY_NOTIFICATION_TIME_DAILY));
                 findPreference(KEY_NOTIFICATION_TIME_DAILY).setSummary(formatTime(time));
-
-                String[] minuteAndTime = time.split(":");
-                int hour = Integer.parseInt(minuteAndTime[0]);
-                int minute = Integer.parseInt(minuteAndTime[1]);
-
-                DailyNotificationJob.schedule(hour, minute);
-
+                DailyNotificationJob.schedule();
                 break;
             case KEY_NOTIFICATION_TIME_WEEKLY:
                 String index = String.valueOf(SettingsSharedPreferences.getInstance().get(KEY_NOTIFICATION_TIME_WEEKLY));
                 findPreference(KEY_NOTIFICATION_TIME_WEEKLY).setSummary(weeks[Integer.parseInt(index)]);
             case KEY_NOTIFICATION_SWITCH_DAILY:
-                //todo check if true
-                cancelTask(DailyNotificationJob.TAG);
+                if (sharedPreferences.getBoolean(key, false)) {
+                    DailyNotificationJob.schedule();
+                } else {
+                    cancelTask(KEY_NOTIFICATION_SWITCH_DAILY);
+                }
+                break;
+            case KEY_NOTIFICATION_SWITCH_WEEKLY:
+                if (sharedPreferences.getBoolean(key, false)) {
+                    WeeklyNotificationJob.schedule();
+                } else {
+                    cancelTask(KEY_NOTIFICATION_SWITCH_DAILY);
+                }
+                break;
+            case KEY_NOTIFICATION_SWITCH_MONTHLY:
+                if (sharedPreferences.getBoolean(key, false)) {
+                    MonthlyNotificationJob.schedule();
+                } else {
+                    cancelTask(KEY_NOTIFICATION_SWITCH_MONTHLY);
+                }
                 break;
         }
     }
@@ -185,122 +193,6 @@ public class SettingsFragment extends PreferenceFragment implements Preference.O
         return time;
     }
 
-    private class CustomTimePickerDialog extends TimePickerDialog {
-        private final String dialogTitle = getContext().getString(R.string.select_time);
-
-        CustomTimePickerDialog(Context context, OnTimeSetListener callBack, int hour, int minute) {
-            super(context, android.R.style.Theme_Holo_Light_Dialog, callBack, hour, minute, DateFormat.is24HourFormat(context));
-            setTitle(dialogTitle);
-            fixSpinner(context, hour, minute, DateFormat.is24HourFormat(context));
-
-            Window window = getWindow();
-            if (window != null) {
-                window.setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
-            }
-        }
-
-        public void setTitle(CharSequence title) {
-            super.setTitle(dialogTitle);
-        }
-
-        /**
-         * Workaround for this bug: https://code.google.com/p/android/issues/detail?id=222208
-         * In Android 7.0 Nougat, spinner mode for the TimePicker in TimePickerDialog is
-         * incorrectly displayed as clock, even when the theme specifies otherwise.
-         * <p>
-         * Source: https://gist.github.com/jeffdgr8/6bc5f990bf0c13a7334ce385d482af9f
-         */
-        @SuppressWarnings("deprecation")
-        private void fixSpinner(Context context, int hourOfDay, int minute, boolean is24HourView) {
-            // android:timePickerMode spinner and clock began in Lollipop
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                try {
-                    // Get the theme's android:timePickerMode
-                    final int MODE_SPINNER = 2;
-                    Class<?> styleableClass = Class.forName("com.android.internal.R$styleable");
-                    Field timePickerStyleableField = styleableClass.getField("TimePicker");
-                    int[] timePickerStyleable = (int[]) timePickerStyleableField.get(null);
-                    final TypedArray a = context.obtainStyledAttributes(null, timePickerStyleable,
-                            android.R.attr.timePickerStyle, 0);
-                    Field timePickerModeStyleableField = styleableClass.getField("TimePicker_timePickerMode");
-                    int timePickerModeStyleable = timePickerModeStyleableField.getInt(null);
-                    final int mode = a.getInt(timePickerModeStyleable, MODE_SPINNER);
-                    a.recycle();
-
-                    if (mode == MODE_SPINNER) {
-                        Field field = findField(TimePickerDialog.class, TimePicker.class, "mTimePicker");
-                        if (field == null) {
-                            Timber.e("Reflection failed: Couldn't find field 'mTimePicker'");
-                            return;
-                        }
-
-                        TimePicker timePicker = (TimePicker) field.get(this);
-                        Class<?> delegateClass = Class.forName("android.widget.TimePicker$TimePickerDelegate");
-                        Field delegateField = findField(TimePicker.class, delegateClass, "mDelegate");
-
-                        if (delegateField == null) {
-                            Timber.e("Reflection failed: Couldn't find field 'mDelegate'");
-                            return;
-                        }
-                        Object delegate = delegateField.get(timePicker);
-
-                        Class<?> spinnerDelegateClass;
-                        if (Build.VERSION.SDK_INT != Build.VERSION_CODES.LOLLIPOP) {
-                            spinnerDelegateClass = Class.forName("android.widget.TimePickerSpinnerDelegate");
-                        } else {
-                            // TimePickerSpinnerDelegate was initially misnamed in API 21!
-                            spinnerDelegateClass = Class.forName("android.widget.TimePickerClockDelegate");
-                        }
-
-                        // In 7.0 Nougat for some reason the timePickerMode is ignored and the
-                        // delegate is TimePickerClockDelegate
-                        if (delegate.getClass() != spinnerDelegateClass) {
-                            delegateField.set(timePicker, null); // throw out the TimePickerClockDelegate!
-                            timePicker.removeAllViews(); // remove the TimePickerClockDelegate views
-                            Constructor spinnerDelegateConstructor = spinnerDelegateClass
-                                    .getConstructor(TimePicker.class, Context.class,
-                                            AttributeSet.class, int.class, int.class);
-                            spinnerDelegateConstructor.setAccessible(true);
-
-                            // Instantiate a TimePickerSpinnerDelegate
-                            delegate = spinnerDelegateConstructor.newInstance(timePicker, context,
-                                    null, android.R.attr.timePickerStyle, 0);
-
-                            // set the TimePicker.mDelegate to the spinner delegate
-                            delegateField.set(timePicker, delegate);
-
-                            // Set up the TimePicker again, with the TimePickerSpinnerDelegate
-                            timePicker.setIs24HourView(is24HourView);
-                            timePicker.setCurrentHour(hourOfDay);
-                            timePicker.setCurrentMinute(minute);
-                            timePicker.setOnTimeChangedListener(this);
-                        }
-                    }
-                } catch (Exception e) {
-                    throw new RuntimeException(e);
-                }
-            }
-        }
-
-        private Field findField(Class objectClass, Class fieldClass, String expectedName) {
-            try {
-                Field field = objectClass.getDeclaredField(expectedName);
-                field.setAccessible(true);
-                return field;
-            } catch (NoSuchFieldException e) {
-                Timber.i(e); // ignore
-            }
-
-            // search for it if it wasn't found under the expected ivar name
-            for (Field searchField : objectClass.getDeclaredFields()) {
-                if (searchField.getType() == fieldClass) {
-                    searchField.setAccessible(true);
-                    return searchField;
-                }
-            }
-            return null;
-        }
-    }
 
     @Override
     public void onDestroy() {
