+++ collect_app/src/main/res/layouts/fieldsight/layout/activity_fieldsight_settings.xml
@@ -5,8 +5,10 @@
     android:layout_height="match_parent"
     android:orientation="vertical">
 
+    <include layout="@layout/toolbar" />
 
     <View
+        android:layout_marginTop="?actionBarSize"
         android:id="@+id/content"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
